added

+++ src/components/headers/Header.js
@@ -35,7 +35,7 @@ const Header = () => {
   const [accountStatus, setAccountStatus] = useState()
   const [showStatus, setShowStatus] = useState(false)
   const [category, setCategory] = useState([])
-  const [mobilView,setMobileView] = useState(false)
+  const [mobilView, setMobileView] = useState(false)
   const isAdmin = useSelector(state => state.User.user_role);
   const [language, setLanguage] = useState(); // Default language is English
   const location = useLocation()
@@ -48,7 +48,7 @@ const Header = () => {
   }, [localStorage.getItem('isLoggedIn')])
   useEffect(() => {
     setLanguage(localStorage.getItem("lan"))
-}, [localStorage])
+  }, [localStorage])
   useEffect(() => {
     dispatch(get_category()).then((res) => {
       // console.log(res)
@@ -106,19 +106,23 @@ const Header = () => {
   }, [screenSize])
   function logoutHandler() {
     setIsOpen(false)
+    if(mobilView){
+    clossAll()}
+    //  setShowMenu(false) 
+    localStorage.clear();
+    localStorage.setItem("isLoggedIn", false);
+    localStorage.setItem("lan", "hi");
+    localStorage.setItem("isModal", false);
     dispatch(logout()).then((res) => {
-      localStorage.clear()
-      localStorage.setItem("isLoggedIn", false)
-      window.location = '/'
-      // window.location = '/login'
-      setisLoggedIn(false)
-    })
+      setisLoggedIn(false);
+      window.location = '/';
+    });
   }
   const handleLanguageChange = (event) => {
     setLanguage(event.target.value);
     localStorage.setItem("lan", event.target.value)
     window.location = location.pathname
-};
+  };
 
 
 
@@ -139,12 +143,24 @@ const Header = () => {
   }
   const handleSettings = () => {
     // Handle settings logic
+    if(mobilView){
+      clossAll()}
     setIsOpen(false)
     navigate('/update-profile')
   };
+  const dashboard = ()=>{
+    setIsOpen(false)
+    if(mobilView){
+      clossAll()}
+    navigate('/dashboard')
+  }
   const handleMenus = () => {
     setShowMenu(!showMenu);
     setIsOpen(false)
+  }
+  function clossAll(){
+    setIsOpen(false)
+    setShowMenu(false)
   }
   return (
     <>
@@ -155,79 +171,80 @@ const Header = () => {
           <TopBar />
           <div className='bg-white'>
             <div className='container py-2'>
-              <header  className={`${showMenu && mobilView ? 'container fixed left-0 top-0 z-[1035] h-screen w-60 -translate-x-full overflow-hidden bg-white shadow-[0_4px_12px_0_rgba(0,0,0,0.07),_0_2px_4px_rgba(0,0,0,0.05)] translate-x-0 dark:bg-zinc-800' : ''} md:h-auto md:grid  md:place-items-center  md:mb-0 md:grid-cols-9`}>
-             
-                  <div className={`${mobilView && !showMenu ? "grid grid-cols-7": ''}`}>
-                <div className='col-span-2 md:col-span-1 mr-3 md:mr-0 '>
-                  <div> <img className='  md-w-auto' onClick={() => navigate('/')} src={NEW_Sourcify} alt="logo" /></div>
-                </div>
-                {
-                  mobilView && showMenu&& <div className='mt-3 h-5  flex  items-center header_text '>
-                  <label htmlFor="language">Language:</label>
-                  <select id="language" value={language} onChange={handleLanguageChange}>
-                      <option value="en">English</option>
-                      <option value="hi">हिंदी</option>
-                  </select>
-  
-              </div>
-                }
-                {isLoggedIn ?
-                  <>
-                    <div className={`col-span-3 md:hidden ${showMenu ? 'mx-0 order-3' : 'mx-3'} flex   my-3 mr-3`}>
-                      <span className='mr-2 '><TiMessages /></span>  <NavLink to="/messages" className=" h-5  flex  items-center header_text"><p className='header_text' data-translate="hi">Messages</p></NavLink>
-                    </div>
-                    <div
-                      onClick={() => setIsOpen(!isOpen)}
-                      className="inline-flex md:hidden col-span-1 h-[70%] my-2 mr-8 md:justify-center items-center  w-full "
-                    >
-                      <img src={profile} />
-                      {/* <span>{userName}</span> */}
-                      <svg
-                        className=" w-[10px]   mr-2"
-                        fill="none"
-                        stroke="currentColor"
-                        viewBox="0 0 24 24"
-                        xmlns="http://www.w3.org/2000/svg"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          strokeWidth={2}
-                          d="M19 9l-7 7-7-7"
-                        />
-                      </svg>
-                    </div>
+              <header className={`${showMenu && mobilView ? 'container fixed left-0 top-0 z-[1035] h-screen w-60 -translate-x-full overflow-hidden bg-white shadow-[0_4px_12px_0_rgba(0,0,0,0.07),_0_2px_4px_rgba(0,0,0,0.05)] translate-x-0 dark:bg-zinc-800' : ''} md:h-auto md:grid  md:place-items-center  md:mb-0 md:grid-cols-9`}>
 
-                  </>
-                  : <div className={ ` ${mobilView && showMenu ?"ml-0 my-3":"ml-8 "} md:hidden grid grid-cols-4 col-span-4`} >
-                    <Link className={`flex   ${mobilView && showMenu ?"col-span-4 mb-2 " : '  col-span-2  my-3 mr-3'}  `} to='/register'><span className='w-auto items-center  bold mr-1'><img src={edit_icon} /></span> <p className='header_text'data-translate="hi">Register</p>  </Link>
-                    <Link className={`flex   ${mobilView && showMenu ?"col-span-4 " : ' col-span-2 my-3 mr-3 '} `} to='/login'><span className='w-auto  items-center bold mr-1'><img src={profile} /></span> <p className='header_text'data-translate="hi"> Login </p> </Link>
-                  </div>}
+                <div className={`${mobilView && !showMenu ? "grid grid-cols-7" : ''}`}>
+                  <div className='col-span-2 md:col-span-1 mr-3 md:mr-0 '>
+                    <div> <img className='  md-w-auto' onClick={() => navigate('/')} src={NEW_Sourcify} alt="logo" /></div>
                   </div>
-                {!showMenu?
-                <div className='absolute col-span-1 cursor-pointer text-2xl right-3 top-[2rem] md:hidden'><MenuOutlined  onClick={() => { handleMenus() }} /></div>
-                :
-                <div className='absolute col-span-1 cursor-pointer text-2xl right-3 top-[2rem] md:hidden'>< RxCross2 onClick={() => { handleMenus() }} /></div>
+                  {
+                    mobilView && showMenu && <div className='mt-3 h-5  flex  items-center header_text '>
+                      <label htmlFor="language">Language:</label>
+                      <select id="language" value={language} onChange={handleLanguageChange}>
+                        <option value="en">English</option>
+                        <option value="hi">हिंदी</option>
+                      </select>
 
+                    </div>
                   }
+                  {isLoggedIn ?
+                    <>
+                      <div className={`col-span-3 md:hidden ${showMenu ? 'mx-0 order-3' : 'mx-3'} flex   my-3 mr-3`}>
+                        <span className='mr-2 '><TiMessages /></span>  <NavLink to="/messages" className=" h-5  flex  items-center header_text"><p className='header_text' data-translate="hi">Messages</p></NavLink>
+                      </div>
+                      <div
+                        onClick={() => setIsOpen(!isOpen)}
+                        className="inline-flex md:hidden col-span-1 h-[70%] my-2 mr-8 md:justify-center items-center  w-full "
+                      >
+                        <img src={profile} />
+                        {/* <span>{userName}</span> */}
+                        <svg
+                          className=" w-[10px]   mr-2"
+                          fill="none"
+                          stroke="currentColor"
+                          viewBox="0 0 24 24"
+                          xmlns="http://www.w3.org/2000/svg"
+                        >
+                          <path
+                            strokeLinecap="round"
+                            strokeLinejoin="round"
+                            strokeWidth={2}
+                            d="M19 9l-7 7-7-7"
+                          />
+                        </svg>
+                      </div>
+
+                    </>
+                    : <div className={` ${mobilView && showMenu ? "ml-0 my-3" : "ml-8 "} md:hidden grid grid-cols-4 col-span-4`} >
+
+                      <Link onClick={() => { setShowMenu(false) }} className={`flex   ${mobilView && showMenu ? "col-span-4 mb-2 " : '  col-span-2  my-3 mr-3'}  `} to='/register'><span className='w-auto items-center  bold mr-1'><img src={edit_icon} /></span> <p className='header_text' data-translate="hi">Register</p>  </Link>
+                      <Link onClick={() => { setShowMenu(false) }} className={`flex   ${mobilView && showMenu ? "col-span-4 " : ' col-span-2 my-3 mr-3 '} `} to='/login'><span className='w-auto  items-center bold mr-1'><img src={profile} /></span> <p className='header_text' data-translate="hi"> Login </p> </Link>
+                    </div>}
+                </div>
+                {!showMenu ?
+                  <div className='absolute col-span-1 cursor-pointer text-2xl right-3 top-[2rem] md:hidden'><MenuOutlined onClick={() => { handleMenus() }} /></div>
+                  :
+                  <div className='absolute col-span-1 cursor-pointer text-2xl right-3 top-[2rem] md:hidden'>< RxCross2 onClick={() => { handleMenus() }} /></div>
+
+                }
                 {showMenu && <>
                   <div className='md:col-span-3 col-span-7 md:grid grid-cols-1 md:grid-cols-3  md:items-center '>
 
-                    <NavLink className='header_text flex md:justify-center cursor-pointer md:mb-0 mb-3 md:ml-2' to={'/contractors'}data-translate="hi" >Find Contractors</NavLink>
+                    <NavLink onClick={() => {mobilView && clossAll() }} className='header_text flex md:justify-center cursor-pointer md:mb-0 mb-3 md:ml-2' to={'/contractors'} data-translate="hi" >Find Contractors</NavLink>
 
-                    <div  className='header_text flex md:justify-center cursor-pointer md:mb-0 mb-3 '>
-                    <NavLink to={'/project_list'}data-translate="hi">Find Projects</NavLink>
+                    <div className='header_text flex md:justify-center cursor-pointer md:mb-0 mb-3 '>
+                      <NavLink onClick={() => {mobilView && clossAll() }} to={'/project_list'} data-translate="hi">Find Projects</NavLink>
                     </div>
-                    <div  className='header_text flex md:justify-center cursor-pointer md:mb-0 mb-3 '>
-                    <NavLink className='header_text flex justify-center cursor-pointer md:mb-0 mb-3' to={'/SourcifyWork'}data-translate="hi">How it works</NavLink>
+                    <div className='header_text flex md:justify-center cursor-pointer md:mb-0 mb-3 '>
+                      <NavLink onClick={() => {mobilView && clossAll() }} className='header_text flex justify-center cursor-pointer md:mb-0 mb-3' to={'/SourcifyWork'} data-translate="hi">How it works</NavLink>
                     </div>
 
-                   
+
                   </div>
                   {showMenu && <>
                     <div className='grid  col-span-7  md:col-span-2 w-full md:mb-0 mb-3 ' >
                       <form onSubmit={submitHandler} className="w-full">
-                        <Input onChange={inputHandler} placeholder='Search for contractors or projects ' className='input_radius w-full' suffix={<img src={search_icon} />} />
+                        <Input onChange={inputHandler} placeholder='Search for contractors or projects ' className='input_radius w-full' suffix={<img onClick={submitHandler} src={search_icon} />} />
                       </form>
                     </div>
                   </>
@@ -236,16 +253,16 @@ const Header = () => {
                     {isLoggedIn ?
                       <>
 
-                        <NavLink to="/messages" className="hidden  md:flex items-baseline mr-3 header_text">
+                        <NavLink onClick={() => {mobilView && clossAll() }} to="/messages" className="hidden  md:flex items-baseline mr-3 header_text">
                           <span className='mr-2'><TiMessages /></span>
-                          <p className='header_text'data-translate="hi">Messages</p>
+                          <p className='header_text' data-translate="hi">Messages</p>
                         </NavLink>
 
                         <button
                           onClick={() => setIsOpen(!isOpen)}
                           className="hidden md:flex items-center justify-center w-full"
                         >
-                          <img src={profile}  />
+                          <img src={profile} />
                           <svg
                             className=" w-[10px] "
                             fill="none"
@@ -269,14 +286,14 @@ const Header = () => {
                         <div className='flex items-center mr-3 hidden md:inline-block' to='/register'>
                           <span className='flex items-center'>
                             <img src={edit_icon} className='mr-1' />
-                            <NavLink to='/register' className='header_text'data-translate="hi">Register</NavLink>
+                            <NavLink onClick={() => {mobilView && clossAll() }} to='/register' className='header_text' data-translate="hi">Register</NavLink>
                           </span>
                         </div>
 
                         <div className='flex items-center mr-3 hidden md:inline-block' >
                           <span className='flex items-center'>
                             <img src={profile} className='mr-1' />
-                            <NavLink to='/login' className='header_text'data-translate="hi">Login</NavLink>
+                            <NavLink onClick={() => {mobilView && setShowMenu(false) }} to='/login' className='header_text' data-translate="hi">Login</NavLink>
                           </span>
                         </div>
 
@@ -295,7 +312,7 @@ const Header = () => {
                         Post Project
                       </div>
                       {
-                        mobilView && <SubHeader/>
+                        mobilView && <SubHeader />
                       }
                     </div>
                   </div>
@@ -304,10 +321,10 @@ const Header = () => {
                   <div className="absolute z-[999] top-[10%] md:top-[88px] w-[210px]  left-[34%]  md:left-[66%] bg-[white]  w-60 mt-2 origin-top-right rounded-md shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none" style={{ overflowWrap: 'break-word' }}>
                     <div className="py-1" style={{ overflowWrap: 'break-word' }}>
                       {showStatus && <div className={`px-4 py-2`}>
-                        <span className='cursor-pointer' onClick={() => { accountStatus === 0 && navigate('/dashboard') }}>{`${showMenu && mobilView ?'Status' :"Account status -"} `} </span>
+                        <span className='cursor-pointer' onClick={() => { accountStatus === 0 && navigate('/dashboard') }}>{`${showMenu && mobilView ? 'Status' : "Account status -"} `} </span>
                         <span>{accountStatus === 0 ? <Tag color="green">Live</Tag> : accountStatus === 1 ? <Tag color="yellow">Under Review</Tag> : <Tag color="volcano">Blocked</Tag>}</span>
                       </div>}
-                      <div className='px-4 py-2' onClick={() => { navigate('/dashboard') }}>Dashboard</div>
+                      <div className='px-4 py-2' onClick={() => {dashboard()  }}>Dashboard</div>
                       <button
                         onClick={handleSettings}
                         className="block px-4 py-2 text-sm w-full text-left"
@@ -327,7 +344,7 @@ const Header = () => {
                           />
                         </svg>
                         <span data-translate="hi">Update Profile</span>
-                       
+
                       </button>
 
                       <button
@@ -349,20 +366,20 @@ const Header = () => {
                           />
                         </svg>
                         <span data-translate="hi">Logout</span>
-                       
+
                       </button>
 
 
                     </div>
-                   
+
                   </div>
-                  
+
                 )}
-               
+
               </header>
               {
-                  !mobilView &&
-              <SubHeader />}
+                !mobilView &&
+                <SubHeader />}
             </div>
           </div>
         </>
